Move loginPage and privateRoute under auth
