+++ src/domain/app/App.tsx
@@ -13,9 +13,9 @@ import HarborsPage from '../harbors/HarborsPageContainer';
 import IndividualHarborPage from '../individualHarbor/IndividualHarborPageContainer';
 import IndividualCustomerPage from '../individualCustomer/IndividualCustomerPageContainer';
 import OfferPage from '../offer/OfferPageContainer';
-import LoginPage from '../login/LoginPage';
+import LoginPage from '../auth/loginPage/LoginPage';
 import Page from '../page/Page';
-import PrivateRoute from '../privateRoute/PrivateRoute';
+import PrivateRoute from '../auth/privateRoute/PrivateRoute';
 import ApplicationsPage from '../applications/ApplicationsPageContainer';
 import IndividualApplicationPage from '../individualApplication/IndividualApplicationPageContainer';
 import PricingPage from '../pricing/PricingPageContainer';
