+++ src/domain/auth/privateRoute/PrivateRoute.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Route, Redirect, RouteProps, useLocation } from 'react-router-dom';
 
-import authService from '../auth/authService';
+import authService from '../authService';
 
 const PrivateRoute: React.SFC<RouteProps> = props => {
   const location = useLocation();
